+++ app/src/main/java/com/shagalalab/noty/domain/repository/NoteRepository.kt
@@ -6,7 +6,7 @@ import com.shagalalab.noty.data.model.NoteDbModel
 interface NoteRepository {
     fun addNote(note: NoteDbModel)
 
-    fun getNote(id: Int): LiveData<NoteDbModel>
+    fun getNote(id: Long): LiveData<NoteDbModel>
 
     fun getNotes(): LiveData<List<NoteDbModel>>
